+++ ContentPatcher/Framework/Patches/EditImagePatch.cs
@@ -53,10 +53,10 @@ namespace ContentPatcher.Framework.Patches
             this.ToArea = toArea;
             this.PatchMode = patchMode;
             this.Monitor = monitor;
-            if (fromArea != null)
-                this.Contextuals.Add(fromArea.X).Add(fromArea.Y).Add(fromArea.Width).Add(fromArea.Height);
-            if (toArea != null)
-                this.Contextuals.Add(toArea.X).Add(toArea.Y).Add(toArea.Width).Add(toArea.Height);
+
+            this.Contextuals
+                .Add(fromArea)
+                .Add(toArea);
         }
 
         /// <summary>Apply the patch to a loaded asset.</summary>
@@ -64,45 +64,55 @@ namespace ContentPatcher.Framework.Patches
         /// <param name="asset">The asset to edit.</param>
         public override void Edit<T>(IAssetData asset)
         {
+            string errorPrefix = $"Can't apply image patch \"{this.LogName}\" to {this.TargetAsset}";
+
             // validate
             if (typeof(T) != typeof(Texture2D))
             {
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\" to {this.TargetAsset}: this file isn't an image file (found {typeof(T)}).", LogLevel.Warn);
+                this.Monitor.Log($"{errorPrefix}: this file isn't an image file (found {typeof(T)}).", LogLevel.Warn);
                 return;
             }
             if (!this.FromAssetExists())
             {
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\" to {this.TargetAsset}: the {nameof(PatchConfig.FromFile)} file '{this.FromAsset}' doesn't exist.", LogLevel.Warn);
+                this.Monitor.Log($"{errorPrefix}: the {nameof(PatchConfig.FromFile)} file '{this.FromAsset}' doesn't exist.", LogLevel.Warn);
                 return;
             }
 
             // fetch data
-            Texture2D source = this.ContentPack.Load<Texture2D>(this.FromAsset);
-            Rectangle sourceArea = this.FromArea != null ? this.FromArea.ToRectangle() : new Rectangle(0, 0, source.Width, source.Height);
-            Rectangle targetArea = this.ToArea != null ? this.ToArea.ToRectangle() : new Rectangle(0, 0, sourceArea.Width, sourceArea.Height);
             IAssetDataForImage editor = asset.AsImage();
+            Texture2D source = this.ContentPack.Load<Texture2D>(this.FromAsset);
+            if (!this.TryReadArea(this.FromArea, 0, 0, source.Width, source.Height, out Rectangle sourceArea, out string error))
+            {
+                this.Monitor.Log($"{errorPrefix}: the source area is invalid: {error}.", LogLevel.Warn);
+                return;
+            }
+            if (!this.TryReadArea(this.ToArea, 0, 0, sourceArea.Width, sourceArea.Height, out Rectangle targetArea, out error))
+            {
+                this.Monitor.Log($"{errorPrefix}: the target area is invalid: {error}.", LogLevel.Warn);
+                return;
+            }
 
             // validate error conditions
             if (sourceArea.X < 0 || sourceArea.Y < 0 || sourceArea.Width < 0 || sourceArea.Height < 0)
             {
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\": source area (X:{sourceArea.X}, Y:{sourceArea.Y}, Width:{sourceArea.Width}, Height:{sourceArea.Height}) has negative values, which isn't valid.", LogLevel.Error);
+                this.Monitor.Log($"{errorPrefix}: source area (X:{sourceArea.X}, Y:{sourceArea.Y}, Width:{sourceArea.Width}, Height:{sourceArea.Height}) has negative values, which isn't valid.", LogLevel.Error);
                 return;
             }
             if (targetArea.X < 0 || targetArea.Y < 0 || targetArea.Width < 0 || targetArea.Height < 0)
             {
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\": target area (X:{targetArea.X}, Y:{targetArea.Y}, Width:{targetArea.Width}, Height:{targetArea.Height}) has negative values, which isn't valid.", LogLevel.Error);
+                this.Monitor.Log($"{errorPrefix}: target area (X:{targetArea.X}, Y:{targetArea.Y}, Width:{targetArea.Width}, Height:{targetArea.Height}) has negative values, which isn't valid.", LogLevel.Error);
                 return;
             }
             if (targetArea.Right > editor.Data.Width)
             {
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\": target area (X:{targetArea.X}, Y:{targetArea.Y}, Width:{targetArea.Width}, Height:{targetArea.Height}) extends past the right edge of the image (Width:{editor.Data.Width}), which isn't allowed. Patches can only extend the tilesheet downwards.", LogLevel.Error);
+                this.Monitor.Log($"{errorPrefix}: target area (X:{targetArea.X}, Y:{targetArea.Y}, Width:{targetArea.Width}, Height:{targetArea.Height}) extends past the right edge of the image (Width:{editor.Data.Width}), which isn't allowed. Patches can only extend the tilesheet downwards.", LogLevel.Error);
                 return;
             }
             if (sourceArea.Width != targetArea.Width || sourceArea.Height != targetArea.Height)
             {
                 string sourceAreaLabel = this.FromArea != null ? $"{nameof(this.FromArea)}" : "source image";
                 string targetAreaLabel = this.ToArea != null ? $"{nameof(this.ToArea)}" : "target image";
-                this.Monitor.Log($"Can't apply image patch \"{this.LogName}\": {sourceAreaLabel} size (Width:{sourceArea.Width}, Height:{sourceArea.Height}) doesn't match {targetAreaLabel} size (Width:{targetArea.Width}, Height:{targetArea.Height}).", LogLevel.Error);
+                this.Monitor.Log($"{errorPrefix}: {sourceAreaLabel} size (Width:{sourceArea.Width}, Height:{sourceArea.Height}) doesn't match {targetAreaLabel} size (Width:{targetArea.Width}, Height:{targetArea.Height}).", LogLevel.Error);
                 return;
             }