+++ ContentPatcher/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added 'pinned key' feature to the `Random` token, which significantly expands what you can do with it (see readme).
+* Added token support in `FromArea` and `ToArea` fields (thanks to spacechase0!).
 * Fixed incorrect warning when an `EditData` patch references a file that doesn't exist when it should be skipped due to a `HasFile` condition.
 
 ## 1.10.1